refactor(zaino-state): remove background validator and validate-on-read
  The finalised-state background validator did concurrent faulting LMDB reads
  during sync (spent/heights/transparent) and re-derived already-known
  invariants every block, with no consensus value: Layer 1 (per-entry checksum)
  detects bitrot, Layer 2 (merkle/parent/hash<->height/spent<->transparent
  self-consistency) only catches write-path bugs. Neither needs a steady-state
  task on the write-hot path.

  Remove the validator, the validated-prefix machinery (validated_tip /
  validated_set), validate-on-read, the opt-in startup sweeps, and the
  validate_on_start config knob. Relocate the non-validating
  resolve_hash_or_height and the metadata check_schema_version into v1.rs and
  delete validation.rs. Read sites resolve hash<->height without validating;
  the DB is Ready on open.

  Layer 1 checksum primitives (verify / verify_stored / keyed_checksum) are
  retained for migrations and metadata load. Serving-path reads now decode
  without re-verifying; at-rest integrity is delegated to the storage layer.
  This is documented in the module header.
