+++ packages/zaino-state/src/chain_index/finalised_state/db/v1/block_shielded.rs
@@ -165,7 +165,7 @@ impl DbV1 {
         height: Height,
     ) -> Result<SaplingTxList, FinalisedStateError> {
         let validated_height = self
-            .resolve_validated_hash_or_height(HashOrHeight::Height(height.into()))
+            .resolve_hash_or_height(HashOrHeight::Height(height.into()))
             .await?;
         let height_bytes = validated_height.to_bytes()?;
 
@@ -205,8 +205,6 @@ impl DbV1 {
                 "invalid block range: end < start".to_string(),
             ));
         }
-
-        self.validate_block_range(start, end).await?;
         let start_bytes = start.to_bytes()?;
         let end_bytes = end.to_bytes()?;
 
@@ -335,7 +333,7 @@ impl DbV1 {
         height: Height,
     ) -> Result<OrchardTxList, FinalisedStateError> {
         let validated_height = self
-            .resolve_validated_hash_or_height(HashOrHeight::Height(height.into()))
+            .resolve_hash_or_height(HashOrHeight::Height(height.into()))
             .await?;
         let height_bytes = validated_height.to_bytes()?;
 
@@ -375,8 +373,6 @@ impl DbV1 {
                 "invalid block range: end < start".to_string(),
             ));
         }
-
-        self.validate_block_range(start, end).await?;
         let start_bytes = start.to_bytes()?;
         let end_bytes = end.to_bytes()?;
 
@@ -417,7 +413,7 @@ impl DbV1 {
         height: Height,
     ) -> Result<CommitmentTreeData, FinalisedStateError> {
         let validated_height = self
-            .resolve_validated_hash_or_height(HashOrHeight::Height(height.into()))
+            .resolve_hash_or_height(HashOrHeight::Height(height.into()))
             .await?;
         let height_bytes = validated_height.to_bytes()?;
 
@@ -458,8 +454,6 @@ impl DbV1 {
                 "invalid block range: end < start".to_string(),
             ));
         }
-
-        self.validate_block_range(start, end).await?;
         let start_bytes = start.to_bytes()?;
         let end_bytes = end.to_bytes()?;
