+++ packages/zaino-state/src/chain_index/finalised_state/db/v1/block_transparent.rs
@@ -147,7 +147,7 @@ impl DbV1 {
         height: Height,
     ) -> Result<TransparentTxList, FinalisedStateError> {
         let validated_height = self
-            .resolve_validated_hash_or_height(HashOrHeight::Height(height.into()))
+            .resolve_hash_or_height(HashOrHeight::Height(height.into()))
             .await?;
         let height_bytes = validated_height.to_bytes()?;
 
@@ -189,8 +189,6 @@ impl DbV1 {
                 "invalid block range: end < start".to_string(),
             ));
         }
-
-        self.validate_block_range(start, end).await?;
         let start_bytes = start.to_bytes()?;
         let end_bytes = end.to_bytes()?;
