+++ packages/zaino-state/src/chain_index/tests/finalised_state/v1.rs
@@ -345,9 +345,6 @@ async fn load_db_backend_from_file() {
         storage: StorageConfig {
             database: DatabaseConfig {
                 path: db_path.clone(),
-                // Keep the opt-in startup verification sweeps covered: this test
-                // loads a pre-built on-disk fixture, exactly the sweep's territory.
-                validate_on_start: true,
                 ..Default::default()
             },
             ..Default::default()
